Add sse class
This is actually an early draft of the class, completely untested and
incomplete. It does not speak the SSE-Protocol yet, but it should make
it easy to send single events to multiple channels, wich in turn can be
used to serve multiple clients in different goroutines, wich is actually
quite clean and already solves all problems, wich are present in the
python implementation.